Removes custom stack to hopefully fix the mem leak.
Instead of using a stack to go from the closing bracket to its
corresponding opening bracket, it uses code similar to that of the code
used to go from an opening bracket to its corresponding closing bracket,
if the current cell == 0. Also fixes a bug which caused a % with a white
background to be printed after the BF program finished printing.